change ResolveSection::fill(...) arguments from size_t to uint32_t

+++ drm/ResolveSection.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <stdint.h>
+#include <cstdint>
 
 namespace cdc {
 
@@ -10,7 +10,7 @@ public:
 	virtual ~ResolveSection() = 0;
 	virtual uint32_t realize(uint32_t sectionId, uint32_t unknown6, uint32_t size, bool& alreadyLoaded); // 0x04
 	virtual uint32_t allocate(uint32_t sectionId, uint32_t sectionSubType, uint32_t unknown6, uint32_t size, bool& alreadyLoaded); // 0x08
-	virtual void fill(uint32_t id, void* src, size_t size, size_t offset) = 0; // 0x18
+	virtual void fill(uint32_t id, void* src, uint32_t size, uint32_t offset) = 0; // 0x18
 	virtual void construct(uint32_t id, void *drmSectionHeaderMaybe) {}; // 0x24
 	virtual void* getWrapped(uint32_t) = 0;     // 0x30
 	virtual void* getBlob(uint32_t);            // 0x34
